Move datasets serialisation
