+++ gammapy/datasets/__init__.py
@@ -1,4 +1,7 @@
+from gammapy.utils.registry import Registry
 from .core import *
 from .map import *
 from .spectrum import *
 from .flux_points import *
+
+DATASETS = Registry([MapDataset, SpectrumDatasetOnOff, FluxPointsDataset])
